extend capabilities of data transfer to accept custom serialized data sources

+++ src/main/java/de/hpi/msc/jschneider/utility/dataTransfer/DataTransferMessages.java
@@ -25,7 +25,7 @@ public class DataTransferMessages
         @NonNull
         private UUID operationId;
         @NonNull
-        private float[] part;
+        private byte[] part;
         @NonNull
         private boolean isLastPart;
     }
